+++ src/Drift.cpp
@@ -713,8 +713,8 @@ void Drift::updateMatrices_()
     size_t id = popIndices_[i];
     std::string paramName = "1/2N_" + bpp::TextTools::toString(id);
 
-    double prevVal = prevParams_.getParameterValue(paramName);
-    double newVal = getParameterValue(paramName);
+    long double prevVal = prevParams_.getParameterValue(paramName);
+    long double newVal = getParameterValue(paramName);
 
     if(newVal != prevVal)
       matrices_[i] *= (newVal / prevVal);
